Implement scanner module.
The basic functionality of the scanner is implemented and tested.
It works for the entire ifj21 language specification.

When called, it reads characters from stdin and returns a token, if
there is a valid sequence of characters on the input, that represent a
lexeme from the ifj21 language.

When it can't construct a valid token, it returns an error token. Right
now, it doesn't have any information about why a token could not be
constructed.

TODO: return more error information

Before requesting a token from the scanner, it needs to be initialized.
After the scanner is no longer needed or the program is about to exit,
the scanner needs to be destroyed.

If a token type requires an attribute, that attribute has a type of
char* (if attribute isn't required, it is NULL).
Integers and numbers are also represented as strings, so they need to be
converted later. But it might be better to convert it in the scanner,
before returning the token.

TODO: decide where to convert integers and numbers