Implement keyword token types.
When scanner finds a lexem that could be an ID or a keyword, it searches
an array of all ifj21 keywords. If the current lexem is in that array,
it is a keyword. The correct keyword token type is set as the token
type. Otherwise the type is set to an id token and the lexem is copied
into the string buffer of the token.

The keyword tokens do not carry any other info, their attribute is
empty.