Change scanner to work with the token attr union.
The token attribute is now a union of string, int and double.

Scanner now converts the integer and number values into actual int and
double types from the parsed string.

String and Id tokens still have the entire parsed string stored in the
attribute.

If a token doesn't require an attribute, it is set to a NULL char
pointer.